requirements.txt
fastapi
uvicorn
jinja2
python-multipart
google-generativeai
httpx
